Server stable?
