+++ Kdux-devtools-plugin/src/main/kotlin/org/mattshoe/shoebox/kduxdevtoolsplugin/server/ServerIntent.kt
@@ -1,0 +1,12 @@
+package org.mattshoe.shoebox.kduxdevtoolsplugin.server
+
+import org.mattshoe.shoebox.org.mattsho.shoebox.devtools.common.UserCommand
+
+sealed interface ServerIntent {
+    data class Command(val command: UserCommand): ServerIntent
+    data class StartDebugging(val storeName: String): ServerIntent
+    data class PauseDebugging(val storeName: String): ServerIntent
+    data class StopDebugging(val storeName: String): ServerIntent
+    data object StopServer: ServerIntent
+    data object StartServer: ServerIntent
+}
