+++ Kdux-devtools-plugin/src/main/kotlin/org/mattshoe/shoebox/kduxdevtoolsplugin/server/ServerState.kt
@@ -1,0 +1,6 @@
+package org.mattshoe.shoebox.kduxdevtoolsplugin.server
+
+sealed interface ServerState {
+    data object Started: ServerState
+    data object Stopped: ServerState
+}
